server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.105.0"
qdrant-client = "^1.7.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
prisma = "^0.11.0"
python-telegram-bot = "^20.7"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
llama-index = "^0.9.15.post2"
google-generativeai = ">=0.3.0"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
